Rename main/./model/person directory to ./musician
Relevant parts are refactored with IDE. Program success without
exception.
