+++ src/main/java/seedu/address/logic/commands/EditCommand.java
@@ -21,8 +21,8 @@ import seedu.address.commons.util.ToStringBuilder;
 import seedu.address.logic.Messages;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
-import seedu.address.model.person.*;
-import seedu.address.model.person.Musician;
+import seedu.address.model.musician.*;
+import seedu.address.model.musician.Musician;
 import seedu.address.model.tag.Tag;
 
 /**
