+++ src/main/java/seedu/address/logic/parser/AddCommandParser.java
@@ -12,11 +12,11 @@ import java.util.stream.Stream;
 
 import seedu.address.logic.commands.AddCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.person.Address;
-import seedu.address.model.person.Email;
-import seedu.address.model.person.Name;
-import seedu.address.model.person.Musician;
-import seedu.address.model.person.Phone;
+import seedu.address.model.musician.Address;
+import seedu.address.model.musician.Email;
+import seedu.address.model.musician.Name;
+import seedu.address.model.musician.Musician;
+import seedu.address.model.musician.Phone;
 import seedu.address.model.tag.Tag;
 
 /**
