+++ src/main/java/seedu/address/model/musician/Musician.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person;
+package seedu.address.model.musician;
 
 import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
