+++ src/main/java/seedu/address/model/musician/NameContainsKeywordsPredicate.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person;
+package seedu.address.model.musician;
 
 import java.util.List;
 import java.util.function.Predicate;
