+++ src/main/java/seedu/address/model/musician/exceptions/PersonNotFoundException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person.exceptions;
+package seedu.address.model.musician.exceptions;
 
 /**
  * Signals that the operation is unable to find the specified musician.
