+++ src/main/java/seedu/address/model/util/SampleDataUtil.java
@@ -6,8 +6,8 @@ import java.util.stream.Collectors;
 
 import seedu.address.model.AddressBook;
 import seedu.address.model.ReadOnlyAddressBook;
-import seedu.address.model.person.*;
-import seedu.address.model.person.Musician;
+import seedu.address.model.musician.*;
+import seedu.address.model.musician.Musician;
 import seedu.address.model.tag.Tag;
 
 /**
