+++ src/main/java/seedu/address/storage/JsonAdaptedPerson.java
@@ -10,11 +10,11 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import seedu.address.commons.exceptions.IllegalValueException;
-import seedu.address.model.person.Address;
-import seedu.address.model.person.Email;
-import seedu.address.model.person.Name;
-import seedu.address.model.person.Musician;
-import seedu.address.model.person.Phone;
+import seedu.address.model.musician.Address;
+import seedu.address.model.musician.Email;
+import seedu.address.model.musician.Name;
+import seedu.address.model.musician.Musician;
+import seedu.address.model.musician.Phone;
 import seedu.address.model.tag.Tag;
 
 /**
