+++ src/test/java/seedu/address/logic/commands/AddCommandIntegrationTest.java
@@ -11,7 +11,7 @@ import seedu.address.logic.Messages;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
-import seedu.address.model.person.Musician;
+import seedu.address.model.musician.Musician;
 import seedu.address.testutil.PersonBuilder;
 
 /**
