+++ src/test/java/seedu/address/logic/commands/AddCommandTest.java
@@ -22,7 +22,7 @@ import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
 import seedu.address.model.ReadOnlyAddressBook;
 import seedu.address.model.ReadOnlyUserPrefs;
-import seedu.address.model.person.Musician;
+import seedu.address.model.musician.Musician;
 import seedu.address.testutil.PersonBuilder;
 
 public class AddCommandTest {
