+++ src/test/java/seedu/address/logic/parser/AddressBookParserTest.java
@@ -23,8 +23,8 @@ import seedu.address.logic.commands.FindCommand;
 import seedu.address.logic.commands.HelpCommand;
 import seedu.address.logic.commands.ListCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.person.Musician;
-import seedu.address.model.person.NameContainsKeywordsPredicate;
+import seedu.address.model.musician.Musician;
+import seedu.address.model.musician.NameContainsKeywordsPredicate;
 import seedu.address.testutil.EditPersonDescriptorBuilder;
 import seedu.address.testutil.PersonBuilder;
 import seedu.address.testutil.PersonUtil;
