+++ src/test/java/seedu/address/testutil/AddressBookBuilder.java
@@ -1,7 +1,7 @@
 package seedu.address.testutil;
 
 import seedu.address.model.AddressBook;
-import seedu.address.model.person.Musician;
+import seedu.address.model.musician.Musician;
 
 /**
  * A utility class to help with building Addressbook objects.
