+++ src/test/java/seedu/address/testutil/EditPersonDescriptorBuilder.java
@@ -5,8 +5,8 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import seedu.address.logic.commands.EditCommand.EditPersonDescriptor;
-import seedu.address.model.person.*;
-import seedu.address.model.person.Musician;
+import seedu.address.model.musician.*;
+import seedu.address.model.musician.Musician;
 import seedu.address.model.tag.Tag;
 
 /**
